deharcode kem secret key length + use snake case + refactor
